requirements.txt
gradio>=4.0.0
openai-whisper>=20231117
openai>=1.0.0
pandas>=1.5.0
yt-dlp>=2023.7.6
PyYAML>=6.0
torch>=2.0.0
torchaudio>=2.0.0